+++ src/git/filefactory.ts
@@ -19,6 +19,6 @@ export class GitFileFactory {
     private static inWorkspace(fileName: string): boolean {
         const uriFileName = Uri.file(fileName);
 
-        return typeof workspace.getWorkspaceFolder(uriFileName) !== "undefined";
+        return !!workspace.getWorkspaceFolder(uriFileName);
     }
 }