+++ src/util/gitcommand.ts
@@ -10,7 +10,7 @@ import { Translation } from "./translation";
 export function getGitCommand(): Promise<string> {
     const gitConfig = workspace.getConfiguration("git");
     const pathCommand = gitConfig.get("path") as string;
-    const promise = new Promise<string>((resolve, reject) => {
+    const promise = new Promise<string>((resolve) => {
         if (!pathCommand) {
             resolve(GIT_COMMAND_IN_PATH);
         }