+++ src/util/property.ts
@@ -1,48 +1,34 @@
-import { Disposable, workspace } from "vscode";
+import { workspace } from "vscode";
 
 export enum Properties {
     CommitUrl = "commitUrl",
     IgnoreWhitespace = "ignoreWhitespace",
     InfoMessageFormat = "infoMessageFormat",
-    InternalHashLength = "internalHashLength",
     IsWebPathPlural = "isWebPathPlural",
     LogLevel = "logLevel",
-    ProgressSpinner = "progressSpinner",
     StatusBarMessageFormat = "statusBarMessageFormat",
     StatusBarMessageNoCommit = "statusBarMessageNoCommit",
     StatusBarPositionPriority = "statusBarPositionPriority",
 }
 
-export class Property {
-    public static get(
-        property:
-            Properties.CommitUrl |
-            Properties.InfoMessageFormat |
-            Properties.StatusBarMessageFormat |
-            Properties.StatusBarMessageNoCommit,
-    ): string;
-    public static get(
-        property:
-            Properties.IgnoreWhitespace |
-            Properties.IsWebPathPlural,
-    ): boolean;
-    public static get(
-        property:
-            Properties.InternalHashLength |
-            Properties.StatusBarPositionPriority,
-    ): number;
-    public static get(
-        property:
-            Properties.LogLevel |
-            Properties.ProgressSpinner,
-    ): string[];
-    public static get(
-        property: Properties,
-    ): string | boolean | number | string[];
-    public static get(
-        property: Properties,
-    ): string | boolean | number | string[] {
-        const properties = workspace.getConfiguration("gitblame");
-        return properties.get(property);
-    }
+export function getProperty(
+    property:
+        Properties.CommitUrl |
+        Properties.InfoMessageFormat |
+        Properties.StatusBarMessageFormat |
+        Properties.StatusBarMessageNoCommit,
+): string;
+export function getProperty(
+    property:
+        Properties.IgnoreWhitespace |
+        Properties.IsWebPathPlural,
+): boolean;
+export function getProperty(
+    property: Properties.StatusBarPositionPriority,
+): number;
+export function getProperty(
+    property: Properties,
+): string | boolean | number {
+    const properties = workspace.getConfiguration("gitblame");
+    return properties.get(property);
 }