+++ src/view.ts
@@ -2,8 +2,7 @@ import { StatusBarAlignment, StatusBarItem, window } from "vscode";
 
 import { GitBlame } from "./git/blame";
 import { IGitCommitInfo } from "./interfaces";
-import { Properties, Property } from "./util/property";
-import { Spinner } from "./util/spinner";
+import { getProperty, Properties } from "./util/property";
 import { TextDecorator } from "./util/textdecorator";
 import { Translation } from "./util/translation";
 
@@ -18,26 +17,20 @@ export class StatusBarView {
 
     private static instance: StatusBarView;
     private readonly statusBarItem: StatusBarItem;
-    private progressInterval: NodeJS.Timer;
-    private readonly spinner: Spinner;
-    private spinnerActive: boolean = false;
+    private readonly spinner: string = "$(sync~spin)";
 
     private constructor() {
         this.statusBarItem = window.createStatusBarItem(
             StatusBarAlignment.Left,
-            Property.get(Properties.StatusBarPositionPriority),
+            getProperty(Properties.StatusBarPositionPriority),
         );
-        this.spinner = new Spinner();
     }
 
     public clear(): void {
-        this.stopProgress();
-        this.setText("", false);
+        this.setText("");
     }
 
     public update(commitInfo: IGitCommitInfo): void {
-        this.stopProgress();
-
         if (commitInfo && !commitInfo.generated) {
             const clickable = !GitBlame.isBlankCommit(commitInfo);
 
@@ -47,49 +40,30 @@ export class StatusBarView {
         }
     }
 
-    public stopProgress(): void {
-        clearInterval(this.progressInterval);
-        this.spinnerActive = false;
-    }
-
     public startProgress(): void {
-        if (this.spinnerActive) {
-            return;
-        }
-
-        this.stopProgress();
-
-        if (this.spinner.updatable()) {
-            this.progressInterval = setInterval(() => {
-                this.setSpinner();
-            }, 100);
-        } else {
-            this.setSpinner();
-        }
-
-        this.spinnerActive = true;
+        this.setText(this.spinner + " " + Translation.do("info.loading_info"));
     }
 
     public dispose(): void {
-        this.stopProgress();
         this.statusBarItem.dispose();
     }
 
-    private setText(text: string, hasCommand: boolean = true): void {
-        this.statusBarItem.text = text
-            ? `$(git-commit) ${text}`
-            : "$(git-commit)";
-        this.statusBarItem.tooltip = hasCommand
-            ? "git blame"
-            : "git blame - " + Translation.do("info.no_info");
-        this.statusBarItem.command = hasCommand ? "gitblame.quickInfo" : "";
-        this.statusBarItem.show();
-    }
+    private setText(text: string, hasCommand: boolean = false): void {
+        if (text) {
+            this.statusBarItem.text = `$(git-commit) ${ text }`;
+        } else {
+            this.statusBarItem.text = "$(git-commit)";
+        }
 
-    private setSpinner(): void {
-        this.setText(
-            this.spinner + " " + Translation.do("info.loading_info"),
-            false,
-        );
+        if (hasCommand) {
+            this.statusBarItem.command = "gitblame.quickInfo";
+            this.statusBarItem.tooltip = "git blame";
+        } else {
+            const noInfo = Translation.do("info.no_info");
+            this.statusBarItem.command = "";
+            this.statusBarItem.tooltip = "git blame - " + noInfo;
+        }
+
+        this.statusBarItem.show();
     }
 }